requirements.txt
asgiref==3.5.2
Django==3.2.16
django-rest-framework==0.1.0
djangorestframework==3.13.1
openai==0.27.2
python-telegram-bot==13.7
python-dotenv==0.19.2
sqlparse==0.4.2
tqdm==4.64.0
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
requests==2.28.1
urllib3==1.26.9